+++ .github/copilot-instructions.md
@@ -1,0 +1,4 @@
+# Copilot Instructions
+
+## Project Guidelines
+- The test expectations need to match the actual task assignment algorithm behavior. The algorithm uses OrderBy(d => d.NextAvailableDate) which makes it deterministic - when developers have the same availability, Alice (alphabetically first) gets assigned first. The expected output files should be updated to reflect Alice's assignments for the first tasks.
