+++ SpreadsheetUtilities.ServiceDefaults/SpreadsheetUtilities.ServiceDefaults.csproj
@@ -1,9 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
+
     <IsAspireSharedProject>true</IsAspireSharedProject>
   </PropertyGroup>
