+++ SpreadsheetUtilities.sln
@@ -10,17 +10,10 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{18066C81-BC94-471E-B150-4BCC8CE900EB}"
 	ProjectSection(SolutionItems) = preProject
 		LICENSE.txt = LICENSE.txt
+		projects.txt = projects.txt
 		README.md = README.md
-		COMPLETION_CHECKLIST.md = COMPLETION_CHECKLIST.md
-		COMPLETION_CONFIRMATION.md = COMPLETION_CONFIRMATION.md
-		EXECUTIVE_SUMMARY.md = EXECUTIVE_SUMMARY.md
-		IMPLEMENTATION_SUMMARY.md = IMPLEMENTATION_SUMMARY.md
-		INDEX.md = INDEX.md
-		PROJECT_STRUCTURE.md = PROJECT_STRUCTURE.md
-		README_IMPLEMENTATION.md = README_IMPLEMENTATION.md
-		SESSION_ARCHITECTURE.md = SESSION_ARCHITECTURE.md
-		USAGE_GUIDE.md = USAGE_GUIDE.md
-		VISUAL_SUMMARY.md = VISUAL_SUMMARY.md
+		tasks.txt = tasks.txt
+		team.txt = team.txt
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{C7DD7E85-59DA-4437-9061-E61FCF5B83E7}"
@@ -28,9 +21,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpreadsheetUtility.Test", "SpreadsheetUtility\SpreadsheetUtility.Test\SpreadsheetUtility.Test.csproj", "{E6184575-AB43-4A1B-BE78-4A87B7FB3631}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SimplifiedUtilityConsole", "SimplifiedUtilityConsole", "{BCB9D4AA-42B9-4422-8A0F-A3A1E9448F47}"
-	ProjectSection(SolutionItems) = preProject
-		SimplifiedUtilityConsole\ReadMe.md = SimplifiedUtilityConsole\ReadMe.md		
-	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SpreadsheetUtility", "SpreadsheetUtility", "{589C6770-9BE3-4742-BE34-345E564C5F7D}"
 EndProject
@@ -91,7 +81,6 @@ Global
 		{69825D0B-BFCB-453B-89A6-9AB30B4D6F6A} = {589C6770-9BE3-4742-BE34-345E564C5F7D}
 		{E6184575-AB43-4A1B-BE78-4A87B7FB3631} = {C7DD7E85-59DA-4437-9061-E61FCF5B83E7}
 		{1118E5C9-FCF2-474A-9355-5DD1CD7D3AF1} = {BCB9D4AA-42B9-4422-8A0F-A3A1E9448F47}
-		{B0E8E23F-7945-4AEA-9344-E84F3D8B3370} = {18066C81-BC94-471E-B150-4BCC8CE900EB}
 		{5F9B3108-25A8-46CF-B012-4C81DDC7A492} = {589C6770-9BE3-4742-BE34-345E564C5F7D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
