+++ SpreadsheetUtility/SpreadsheetUtility.Test/SpreadsheetUtility.Test.csproj
@@ -1,9 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
+
 
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
