+++ SpreadsheetUtility/SpreadsheetUtility.UI.Console/SpreadsheetUtility.UI.Console.csproj
@@ -2,9 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
+
   </PropertyGroup>
 
   <ItemGroup>
